# [core] Mark used offline region resources in batches