Marking offline region resources "used" is on the hot code path for offline region download.
Previously, the engine attempted to update the database separetely for each resource and it caused a huge performance impact, especially when some tiles were already loaded.

Now, offline region resources are marked in batches (200 items at once).

Consider the time mesaurements of the `mbgl-offline` tool execution (Linux x64, release build):

`time ./mbgl-offline --north 41.4664 --west 2.0407 --south 41.2724 --east 2.2680 --output barcelona.db`

without patch:
 
1st time:
```
    real      0m9,505s
    user      0m3,315s
    sys       0m0,378s
 ```
    
2nd time (tiles are already loaded):
 ```
    real      0m9,480s
    user      0m0,403s
    sys       0m0,645s
 ```

with the patch:
    
1st time:
```
    real      0m7,399s
    user      0m2,994s
    sys       0m0,384s
 ```
    
2nd time (tiles are already loaded):
```
    real      0m0,149s
    user      0m0,059s
    sys       0m0,044s
```

So, the proposed changes make the repeated load **~64** times faster :tada: