@alexshalamov 2019-08-29
@LukasPaczos would you be able to test this PR for Android SDK?