@LukasPaczos 2019-08-29
The improvements are amazing!

A small offline region that takes about 8s to download:

|  | Previously | Now |
| ------------- | ------------- | ------------- |
| First re-download in the current process | 8s | 1.3s  |
| Subsequent re-downloads in the same process | 4.5s | 0.7s |

What's interesting, first re-download in the app process (for example after restarting the application when tiles are already in place) takes significantly longer than subsequent ones.