@andrewharvey 2019-08-29
Wow I didn't realise you could call mbgl-offline twice on the same output and it would just update, I always thought it would just replace it and re-download everything. Could we update this in the documentation? PS. I can't find any of the offline sideload documentation anymore.