@LukasPaczos 2019-08-29
A region with `5001` resources that originally takes ~47s to download:

Testing different batch sizes https://github.com/mapbox/mapbox-gl-native/pull/15521/files#diff-72890ff2093a9bfa6bfe3a9e8955d6dfR361:
#### For batches of `200` resources (as in the PR)

|  | Previously | Now |
| ------------- | ------------- | ------------- |
| First re-download in the current process | 41s | 2.7s  |
| Subsequent re-downloads in the same process | 38s | 1.3s |

#### For batches of `1000` resources

|  | Previously | Now |
| ------------- | ------------- | ------------- |
| First re-download in the current process | 41s | 1.6s  |
| Subsequent re-downloads in the same process | 38s | 1.0s |