@pozdnyakov 2019-08-29
We could make batch size adaptive, i.e. depend on the total resources count, I'll investigate it and implement in a follow-up PR.